fix(keys): bug fix comment.id wrong type of data structure was sent with the response
when trying to access the comments.id in the CommentsList.js in the client dir we unsuccessfully attend to retrieve the comment.id because the server's response to the client was an empty array, hence the client did not throw an error. They key was invalid for every rendering cicle of the li since there was no property called ket since there resposne had no object in it.